% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggNaruto-package.R
\docType{package}
\name{ggNaruto-package}
\alias{ggNaruto}
\alias{ggNaruto-package}
\title{ggNaruto: Naruto Color Palettes}
\description{
Colour palettes inspired by Naruto, ported to R 
    for your enjoy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NotebookOFXiaoMing/ggNaruto}
  \item Report bugs at \url{https://github.com/NotebookOFXiaoMing/ggNaruto/issues}
}

}
\keyword{internal}
