% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_naruto_c}
\alias{scale_colour_naruto_c}
\title{Continuous scales to use for ggplot2}
\usage{
scale_colour_naruto_c(name, direction = 1, ...)
}
\arguments{
\item{name}{Name of desired palette.}

\item{direction}{Either '1' or '-1'. If '-1' the palette will be reversed.}

\item{...}{Arguments to pass on to `ggplot2::scale_colour_gradientn` or
`ggplot2::scale_fill_gradientn()`}
}
\value{
A `ScaleContinuous` object that can be added to a `ggplot` object
}
\description{
These functions provide the option to use Naruto colours inside continuous
palettes with the `ggplot2` package.
}
