% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_naruto_d}
\alias{scale_colour_naruto_d}
\alias{scale_fill_naruto_d}
\title{Discrete scales to use for ggplot2}
\usage{
scale_colour_naruto_d(name, direction = 1, ...)

scale_fill_naruto_d(name, direction = 1, ...)
}
\arguments{
\item{name}{Name of desired palette.}

\item{direction}{Either '1' or '-1'. If '-1' the palette will be reversed.}

\item{...}{additional arguments to pass to discrete_scale}
}
\value{

}
\description{
These functions provide the option to use \code{Naruto} colours inside discrete
palettes with the `ggplot2` package.
}
