% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{naruto_palette}
\alias{naruto_palette}
\title{Naruto palette generator}
\usage{
naruto_palette(name, n, direction = 1, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette.}

\item{n}{Number of colors desired.}

\item{direction}{Either '1' or '-1'. If '-1' the palette will be reversed.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
These are some color palettes from Naruto.
}
\examples{
naruto_palette('Team7')
naruto_palette('Team7',n=8)
\dontrun{
naruto_palette("Jiraiya")
}
}
\keyword{colors}
