% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naruto_palettes}
\alias{naruto_palettes}
\title{Complete list of available naruto palettes}
\format{
An object of class \code{list} of length 2.
}
\usage{
naruto_palettes
}
\description{
A list of all colour palettes to choose from, and preview directly in the console.
}
\keyword{datasets}
