% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_naruto_c}
\alias{scale_fill_naruto_c}
\title{Continuous scales to use for ggplot2}
\usage{
scale_fill_naruto_c(name, direction = 1, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Continuous scales to use for ggplot2
}
